refactor: deprecate (root) folder
